Release notes unit tests
