+++ src/components/pageHeader/PageHeader.scss
@@ -7,4 +7,5 @@
   font-weight: 700;
   font-size: 20px;
   line-height: 26px;
+  margin: 0.8rem 0;
 }
